+++ tests/test-wikimedia-commons-edited-videos.sh
@@ -86,15 +86,6 @@ for tool in "${REQUIRED_TOOLS[@]}"; do
     fi
 done
 
-# check if required python scripts exist
-REQUIRED_PYTHON_SCRIPTS=(extract-features.py media-metadata.py create-index.py serve.py)
-for script in "${REQUIRED_PYTHON_SCRIPTS[@]}"; do
-    if [ ! -f "${script}" ]; then
-        echo "$script not found, please run this script from the WISE code directory"
-        exit 1
-    fi
-done
-
 # check if required python packages exist
 REQUIRED_PYTHON_PACKAGES=(torch torchvision torchaudio transformers faiss msclap open_clip)
 for package in "${REQUIRED_PYTHON_PACKAGES[@]}"; do
@@ -127,7 +118,7 @@ fi
 if [ ! -d "${WISE_PROJECT_DIR}" ]; then
     echo "Extracting audio features from videos with ${NUM_WORKERS} workers (takes about 3 min.) ..."
     cd "${WISE_CODE_DIR}"
-    python extract-features.py \
+    python -m wise extract-features \
         "${TEST_DATA_DIR}" \
         --media-include "*.mp4" \
         --shard-maxcount 16 \
@@ -164,7 +155,7 @@ METADATA_TABLE_NAME="metadata-${TEST_ID}"
 RESULT=$(sqlite3 "$METADATA_DB_FILE" "SELECT name FROM sqlite_master WHERE type='table' AND name='$METADATA_TABLE_NAME';")
 if [ "$RESULT" != "$METADATA_TABLE_NAME" ]; then
     echo "Importing metadata from ${TEST_DATA_DIR}/media-metadata.csv (takes few seconds) ..."
-    python3 media-metadata.py import \
+    python3 -m wise media-metadata import \
         --metadata-id "${TEST_ID}" \
         --from-csv "${TEST_DATA_DIR}/media-metadata.csv" \
         --metadata-type "media" \
@@ -208,7 +199,7 @@ fi
 ROW_COUNT=$(sqlite3 "$METADATA_DB_FILE" "SELECT COUNT(*) FROM '$SHOTS_TABLE_NAME';")
 if [ "$ROW_COUNT" -eq 0 ]; then
     # import shots metadata
-    python3 media-metadata.py \
+    python3 -m wise media-metadata \
         import-shots \
         --project-dir "$WISE_PROJECT_DIR" \
         --from-csv "${TEST_DATA_DIR}/shots.csv"
@@ -234,7 +225,7 @@ FEATURE_STORE1="${WISE_PROJECT_DIR}store/${VIDEO_FEATURE_ID1}/"
 if [ ! -d "${FEATURE_STORE1}" ]; then
     cd "${WISE_CODE_DIR}"
     echo "Extracting features from videos (takes about 3 min.) ..."
-    python extract-features.py \
+    python -m wise extract-features \
         --yes \
         --media-include "*.mp4" \
         --shard-maxcount 32 \
@@ -253,7 +244,7 @@ FEATURE_STORE2="${WISE_PROJECT_DIR}store/${VIDEO_FEATURE_ID2}/"
 if [ ! -d "${FEATURE_STORE2}" ]; then
     cd "${WISE_CODE_DIR}"
     echo "Extracting face features from videos (takes about 3 min.) ..."
-    python extract-features.py \
+    python -m wise extract-features \
         --yes \
         --media-include "*.mp4" \
         --shard-maxcount 64 \
@@ -295,7 +286,7 @@ echo "{ \"metadata-${TEST_ID}\": [ \"description\", \"date\", \"source_url\", \"
 if [ ! -f "${VIDEO_INDEX_FILENAME1}" ] || [ ! -f "${VIDEO_INDEX_FILENAME2}" ] || [ ! -f "${AUDIO_INDEX_FILENAME}" ]; then
     echo "Creating search index (takes about 1 min.) ..."
     cd "${WISE_CODE_DIR}"
-    python create-index.py \
+    python -m wise create-index \
            --index-type "${FAISS_INDEX_TYPE}" \
            --fts-config "${FTS_CONFIG_FILE}" \
            --project-dir "$WISE_PROJECT_DIR"
@@ -332,7 +323,7 @@ fi
 
 echo "Starting WISE server on ${HTTP_SERVER_HOST}:${HTTP_SERVER_PORT} ( takes about 1 min.) ..."
 cd "${WISE_CODE_DIR}"
-USE_SHOTS=1 LISTEN_ADDRESS=$HTTP_SERVER_HOST PORT=$HTTP_SERVER_PORT python serve.py \
+USE_SHOTS=1 LISTEN_ADDRESS=$HTTP_SERVER_HOST PORT=$HTTP_SERVER_PORT python -m wise serve \
         --index-type "${FAISS_INDEX_TYPE}" \
         --project-dir "$WISE_PROJECT_DIR" & # to start the server in the background
 SERVER_PID=$!